dbfw_tools/requirements.txt
# Runtime dependencies
google-genai>=1.0.0
Pillow>=10.0.0
tkinterdnd2>=0.4.2
pyautogui>=0.9.54
pygetwindow>=0.0.9
numpy>=1.24.0

# Build dependency (not needed at runtime)
pyinstaller>=6.0.0
